+++ NEWS.md
@@ -1,5 +1,6 @@
 # nycgeo 0.1.0.9000
 
+* Added school district, police precinct, city council, and U.S. Congressional boundaries.
 * Updated `nycgeo::nta_acs_data` to include properly aggregated data from `nycgeo::tract_acs_data`. Also, removed NTAs without population (e.g. parks, cemeteries, airports) from `nycgeo::tract_acs_data`.
 * Fixed bug in `nyc_boundaries()` when filtering by cd.
 * Added tests.